added the blogs separately
