+++ src/data/blogs-data.tsx
@@ -1,0 +1,26 @@
+const blogs = [
+  {
+    title: "Dad: A Villain but a Hero",
+    description:
+      "An inspiring short tale of a boy who eradicated “poverty” for his entire bloodline.",
+    date: "14th Dec, 2023",
+    author: "Yuvi",
+    image:
+      "https://miro.medium.com/v2/resize:fit:1400/format:webp/0*2-QJXvDsy2jmhsLY",
+    tags: ["Philosophy", "Life"],
+    url: "https://medium.com/@meyuviofficial/my-dad-villain-but-a-hero-c9bafd9a804e",
+  },
+  {
+    title: "How to handle Multi-Cluster setup for AKS in your local machine",
+    description:
+      "Setting up the multi-cluster setup for AKS in your local machine.",
+    date: "17th May, 2023",
+    author: "Yuvi",
+    image:
+      "https://miro.medium.com/v2/resize:fit:1400/format:webp/0*u1yLlMvnpMeTOGgu",
+    tags: ["Kubernetes", "AKS"],
+    url: "https://medium.com/@meyuviofficial/how-to-handle-multi-cluster-setup-for-aks-in-your-local-machine-2ff8f9b0f791",
+  },
+];
+
+export default blogs;
